+++ cardano-crypto-class/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for `cardano-crypto-class`
 
+## 2.5.0.1
+
+*
+
 ## 2.5.0.0
 
 * Drop mu2 from `PossessionProofDSIGN`